Mv sortedSizeArrayElement definition to word.h, and test in word_test
